feat: context with atom
