+++ web/package.json
@@ -18,6 +18,7 @@
     "astro": "^2.1.3",
     "astro-icon": "^0.8.0",
     "classnames": "^2.3.2",
+    "nanostores": "^0.8.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-icons": "^4.8.0",
