+++ web/src/components/Cart/store.ts
@@ -1,0 +1,10 @@
+import store from "nanostores";
+
+export type CartItem = {
+  productTitle: string;
+  description: number;
+  quantity: number;
+  unitPrice: number;
+};
+
+export const cartItem = store.map<Record<string, CartItem>>({});
